Fix glance-api deployment in a DCN environment
This fixes an error deploying glance-api in DCN deployments by
deleting the extraneous kolla_config from the GlanceApiInternal
THT. Originally, it was thought the public and internal instances
of the glance-api service required their own kolla JSON files, but
this is not necessary (a single JSON file can be used by both
services). In fact, the volume mounts for the internal service
specified the public service's JSON file. The JSON file intended
for the internal service is not being used.

The problem arose on DCN nodes, which only deploy the internal
service. A JSON file for the internal service was generated, but
the volume mount specified a public service's JSON file, which
doesn't exist on DCN nodes.

This is fixed by deleting the kolla_config section that created
a JSON file for the internal service. By doing that all services
(public and internal) will use the same JSON file, which will
be present on all nodes, including DCN nodes where only the internal
service runs.

Resolves: rhbz#2188477
Change-Id: I2f75b450710a9f36f0a3496c41a9314a74217103