Don't require client cert in horizon
For some unknown reason, Horizon's Apache configuration enables
requiring of client certificate when tls-everywhere is enabled, but
since haproxy is not configured to use a client certificate, it all
breaks horribly.

From what I can say looking at the history of commits, the
SSLVerifyClient parameter was included in tripleo-templates from
the beginning, but with an invalid value of "true", so it had
no effect and everything worked fine. At some point the value
was fixed to correct "require", but looks like that was never
tested, either by the author or by the reviewers, and ever since
then Horizon is broken when tls-everywhere is enabled.

We have two ways of fixing this. Either remove the ssl_verify_client
like this patch is doing, or add the client certificates to the
haproxy configuration. Since no other service is using client
certificates like that, I choose to go with an option that brings
Horizon to the same setup as other services.

Change-Id: I7d3f833cf36d7169a0fbc25d133284e06d3f1468
Resolves: rhbz#2193388