[ffwd3] Prevent renaming nics with specific prefix
On some HW we have nics changing name during system upgrade.
This breaks network and in order to prevent this we allow
specifying list of NIC prefixes to generate udev rules in
order to prevent the rename.

Example would be node with enoX ensY and enpZ interfaces where
we tested on preprod that eno and enp keep renaming. In this
case we would set in system_upgrade.yaml:

parameter_defaults:
  NICsPrefixesToUdev: ['eno','enp']

Co-Authored-By: Sergii Golovatiuk <sgolovat@redhat.com>

Change-Id: I5d7b731a2a284c817a8618a685a27b0a54dfc86b
Related-Bug: rhbz#2142760