Designate: conditionally force enable_new_defaults on
If the DesignateApiPolicies is set (as per enable-secure-rbac.yaml),
designate deployment and operations will fail unless
enforce_new_defaults is set to true.

So far, the alternate solutions aren't great. This is a lightweight
solution and is expected to work except in unusual situations since we
added the DesignateApiPolicies variable in order to support secure
rbac. Other solutions involve adding "yet another variable" or just
another variation of conditionally setting the param to true somewhere
else. This will also function properly once EnforceSecureRbac is set
to true globally when it is supported.

Change-Id: Id0cebaaf91745dd0d9c7cc66c676c6f9fdad413a