Add the possibility to override the OS definition for enforcement.
There is an inline check for the Red Hat version of OpenStack that
ensure that the OS version matches the OpenStack version.

In the latest version of it we need to be able to temporarily change
the os version for some roles.

We implement this by defining a new role variable that will be
propagated to the ansible role responsible for the enforcement.

Note that when the rhsm_enforce_multios is defined, that create an
extra newline, but it doesn't break the resulting ansible playbook as
indentation is respected.

Depends-On: https://review.opendev.org/c/openstack/tripleo-ansible/+/891881
Change-Id: Ie9e789cb6367c7b1f96bf4f8a45aa08caa698fbc