[ffwd3][update] Run OVN external update only once
The ovn_controller update tasks are in external_update but
reused by upgrade process. During the upgrade we run on
all hosts and not just undercloud causing the tasks to get
executed n times over so in case of 500 nodes we would end
up restarting ovn_controller 500 times.

This should be solved in upgrade by simply triggering only
on first controller.

The run_once would not work here due to tripleo_free strategy.

Change-Id: I04117dec68f0892ee59f0b997fb6eb0443218846
Resolves: rhbz#2229761