[ffwd3][update] Improve ovn update
To ensure we get the right config we now use config_patterns.
Also to find ovn config file we use find to pick on all
ovn_controller.json. This should help with upgrade case when
the config is hashed-ovn_controller.json. Lastly if we use
config_patterns we need to pass image override also as an
pattern.

Resolves: rhbz#2243277
Depends-On: I33a5b4d9b6335ae07f84c4f9f0c7629ed6bd6180
Change-Id: Ie729fd31d6a09f661e7c5e7df8435ff2ebdad3b4