Manage QEMU driver cgroups by monolithic libvirtd
We execute monolithic libvirtd in a transient scope, but when the
service unit for it gets restarted by systemd, the qemu driver
controllers' cgroups get rewritten for the transient run-rXXX scopes.

That deactivates the QEMU driver in libvirt as the expected controllers
are no longer there.

Workaround that for newly created Nova instances by delegating the QEMU
cgroup controllers management to libvirt, so that systemd no longer
"sees" its cgroups w/o a proper machined integration in place
(which is when host VMs register in it).

NOTE: Tripleo cannot use systemd-machined for cgrups management
because of upgrade blockers (all VMs not registered in machined get
killed during upgrade). At some point we really need to just
adopt systemd-machined so we can stop fighting against systemd, but
that requires an upgrade process to be written, which is out of scope
of today Tripleo project.

Related: rhbz#2228376
Related: rhbz#2233089

Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>
Change-Id: I1320c6dd01376b0f821f223b6c30a97456b16085